web_client/requirements.txt
Django==2.2
djangorestframework==3.10.3
urllib3==1.24.3
djangorestframework-jwt
django-crispy-forms
